requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
easyocr==1.7.2
exceptiongroup==1.3.1
fastapi==0.109.2
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
ninja==1.13.0
numpy==2.0.2
openai==2.15.0
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
packaging==26.0
pdf2image==1.17.0
pillow==11.3.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.2
pyclipper==1.3.0.post6
pydantic==1.10.26
pydantic_core==2.41.5
pyparsing==3.3.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.14.0
tifffile==2024.8.30
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.8
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.39.0
zipp==3.23.0
